Add app permissions to a bunch of tests that need it (mostly test_new_conversation).
