+++ go/apps/http_api/tests/test_views.py
@@ -5,6 +5,7 @@ class HttpApiTestCase(DjangoGoApplicationTestCase):
     TEST_CONVERSATION_TYPE = u'http_api'
 
     def test_new_conversation(self):
+        self.add_app_permission(u'go.apps.http_api')
         self.assertEqual(len(self.conv_store.list_conversations()), 0)
         response = self.post_new_conversation()
         self.assertEqual(len(self.conv_store.list_conversations()), 1)
