+++ go/apps/opt_out/tests/test_views.py
@@ -5,6 +5,7 @@ class OptOutTestCase(DjangoGoApplicationTestCase):
     TEST_CONVERSATION_TYPE = u'opt_out'
 
     def test_new_conversation(self):
+        self.add_app_permission(u'go.apps.opt_out')
         self.assertEqual(len(self.conv_store.list_conversations()), 0)
         response = self.post_new_conversation()
         self.assertEqual(len(self.conv_store.list_conversations()), 1)
