+++ go/conversation/tests.py
@@ -1,5 +1,3 @@
-import uuid
-
 from django.core.urlresolvers import reverse
 from django.utils.unittest import skip
 
@@ -19,15 +17,6 @@ class ConversationTestCase(VumiGoDjangoTestCase):
         self.setup_api()
         self.setup_user_api()
         self.setup_client()
-
-    def add_app_permission(self, application):
-        permission = self.api.account_store.application_permissions(
-            uuid.uuid4().hex, application=application)
-        permission.save()
-
-        account = self.user_api.get_user_account()
-        account.applications.add(permission)
-        account.save()
 
     def test_get_new_conversation(self):
         self.add_app_permission(u'go.apps.bulk_message')
