+++ go/wizard/tests.py
@@ -13,16 +13,20 @@ class WizardViewsTestCase(VumiGoDjangoTestCase):
         self.setup_client()
 
     def test_get_create_view(self):
+        self.add_app_permission(u'go.apps.bulk_message')
+        self.add_app_permission(u'go.apps.subscription')
         self.declare_tags(u'longcode', 4)
         self.add_tagpool_permission(u'longcode')
         response = self.client.get(reverse('wizard:create'))
         # Check that we have a few conversation types in the response
         self.assertContains(response, 'bulk_message')
         self.assertContains(response, 'subscription')
+        self.assertNotContains(response, 'survey')
         # Check that we have a tagpool/tag in the response
         self.assertContains(response, 'longcode:')
 
     def test_post_create_view_valid(self):
+        self.add_app_permission(u'go.apps.bulk_message')
         self.declare_tags(u'longcode', 4)
         self.add_tagpool_permission(u'longcode')
         self.assertEqual(0, len(self.user_api.active_conversations()))
@@ -41,6 +45,7 @@ class WizardViewsTestCase(VumiGoDjangoTestCase):
             }))
 
     def test_post_create_view_extra_endpoints(self):
+        self.add_app_permission(u'go.apps.wikipedia')
         self.declare_tags(u'longcode', 4)
         self.add_tagpool_permission(u'longcode')
         self.assertEqual(0, len(self.user_api.active_conversations()))
@@ -71,6 +76,7 @@ class WizardViewsTestCase(VumiGoDjangoTestCase):
         self.assertEqual(response.status_code, 200)
 
     def test_post_create_view_invalid_country(self):
+        self.add_app_permission(u'go.apps.bulk_message')
         response = self.client.post(reverse('wizard:create'), {
             'conversation_type': 'bulk_message',
             'name': 'My Conversation',
@@ -82,6 +88,7 @@ class WizardViewsTestCase(VumiGoDjangoTestCase):
         self.assertEqual(response.status_code, 200)
 
     def test_post_create_view_invalid_channel(self):
+        self.add_app_permission(u'go.apps.bulk_message')
         response = self.client.post(reverse('wizard:create'), {
             'conversation_type': 'bulk_message',
             'name': 'My Conversation',
